Implemented StarryPy as a Python package.
Added a setup.py file, with proper hooks for installing with pip. Upon
first run, the user configuration directory is automatically generated,
along with a default config.yaml file and the user plugins directory.
Updated the requirements file to list the proper YAML module. Also
migrated the __main__.py script to starrypy.py, make it more obvious as
the script to be run if executing directly from within the build
directory.

Updated config manager and plugin manager to now respect both paths
for loading plugins.

The default config.yaml is purposefully written in such a way as to be
broken, forcing the user to have to update it before first use.